+++ Application/app/phone-bank/page.tsx
@@ -1,0 +1,5 @@
+import EventTemplateView from "@/app/components/events/EventTemplateView";
+
+export default function PhoneBankPage() {
+  return <EventTemplateView eventType="internal" title="Internal Phone Bank" />;
+}
